+++ momentum/src/scenes/DotPlots/NormalDistribution.tsx
