+++ momentum/src/scenes/Project/styles.css
@@ -1,0 +1,5 @@
+body {
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif;
+}
